--- src/output/summary.txt
@@ -1,5 +1,0 @@
-Based on the provided meeting notes, it appears there was a meeting scheduled for December 25th, 2024, as indicated by the "date" key. However, there is no specific content listed for this meeting, as the "content" array is empty.
-
-To make sense of these meeting notes, we would need additional information or context regarding the purpose, participants, and topics discussed during the meeting. Without the actual content or notes from the meeting, it's difficult to provide a summary or understanding of what transpired.
-
-If you have any further details or if the meeting notes were meant to include additional information, please provide them for a more comprehensive interpretation.
