--- src/screenshare/startScreenshare.ts
@@ -1,65 +1,0 @@
-import { WebDriver } from 'selenium-webdriver';
-
-export async function startScreenshare(driver: WebDriver) {
-  console.log('Starting screen share...');
-
-  await driver.executeScript(`
-    function wait(delayInMS) {
-      return new Promise((resolve) => setTimeout(resolve, delayInMS));
-    }
-
-    function startRecording(stream, lengthInMS) {
-      let recorder = new MediaRecorder(stream);
-      let data = [];
-      
-      recorder.ondataavailable = (event) => data.push(event.data);
-      recorder.start();
-      
-      let stopped = new Promise((resolve, reject) => {
-        recorder.onstop = resolve;
-        recorder.onerror = (event) => reject(event.name);
-      });
-      
-      let recorded = wait(lengthInMS).then(() => {
-        if (recorder.state === "recording") {
-          recorder.stop();
-        }
-      });
-      
-      return Promise.all([stopped, recorded]).then(() => data);
-    }
-
-    window.navigator.mediaDevices.getDisplayMedia({
-      video: { displaySurface: "browser" },
-      audio: true,
-      preferCurrentTab: true
-    }).then(async screenStream => {
-      const audioContext = new AudioContext();
-      const screenAudioStream = audioContext.createMediaStreamSource(screenStream);
-
-      const dest = audioContext.createMediaStreamDestination();
-      screenAudioStream.connect(dest);
-
-      const combinedStream = new MediaStream([
-        ...screenStream.getVideoTracks(),
-        ...dest.stream.getAudioTracks()
-      ]);
-
-      const recordedChunks = await startRecording(combinedStream, 10000);
-      let recordedBlob = new Blob(recordedChunks, { type: "video/webm" });
-
-      const downloadButton = document.createElement("a");
-      downloadButton.href = URL.createObjectURL(recordedBlob);
-      downloadButton.download = "RecordedScreenWithAudio.webm";
-
-      downloadButton.addEventListener('click', () => {
-        window.localStorage.setItem('downloadClicked', 'true');
-      });
-
-      downloadButton.click();
-      screenStream.getTracks().forEach(track => track.stop());
-    });
-  `);
-
-  console.log('Screenshare initialized');
-}
