+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/main.ts","./src/server.ts","./src/drivers/chromedriver.ts","./src/logstojson/savelogstojson.ts","./src/meetproviders/googlemeetautomation.ts","./src/screenshare/startscreenshare.ts","./src/summarizer/summarmeetingnotes.ts"],"version":"5.7.2"}
+{"root":["./src/main.ts","./src/server.ts","./src/drivers/chromedriver.ts"],"version":"5.7.2"}
